refactor: rename file name for consistency
